[change a decimal]

+++ InnovaFunding.Functions/Contract/IDatabaseService.cs
@@ -3,8 +3,8 @@ namespace InnovaFunding.Functions.Contract
 {
     public interface IDatabaseService
     {
-        Task<(double? PriceSales, double? PricePurchase)> GetYesterdayRateAsync(string date);
-        Task InsertRateAsync(string datePublic, double? priceSales, double? pricePurchase);
+        Task<(decimal? PriceSales, decimal? PricePurchase)> GetYesterdayRateAsync(string date);
+        Task InsertRateAsync(string datePublic, decimal? priceSales, decimal? pricePurchase);
         Task LogErrorAsync(string message, string stackTrace);
     }
 }
